+++ src/lotus_json/actors/states/mod.rs
@@ -1,5 +1,6 @@
 // Copyright 2019-2025 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
+
 use super::*;
 mod account_state;
 mod cron_state;
@@ -7,7 +8,6 @@ mod datacap_state;
 mod evm_state;
 mod init_state;
 mod market_state;
-mod methods;
 mod miner_state;
 mod multisig_state;
 mod payment_channel_state;