+++ src/rpc/registry/actors/evm.rs
@@ -5,6 +5,7 @@ use crate::rpc::eth::types::GetStorageAtParams;
 use crate::rpc::registry::methods_reg::{MethodRegistry, register_actor_methods};
 use crate::shim::message::MethodNum;
 use cid::Cid;
+use fil_actors_shared::actor_versions::ActorVersion;
 use fvm_ipld_encoding::RawBytes;
 
 macro_rules! register_evm_version {
@@ -39,15 +40,20 @@ macro_rules! register_evm_version {
     }};
 }
 
-pub(crate) fn register_evm_actor_methods(registry: &mut MethodRegistry, cid: Cid, version: u64) {
+pub(crate) fn register_evm_actor_methods(
+    registry: &mut MethodRegistry,
+    cid: Cid,
+    version: ActorVersion,
+) {
     match version {
-        10 => register_evm_version!(registry, cid, fil_actor_evm_state::v10),
-        11 => register_evm_version!(registry, cid, fil_actor_evm_state::v11),
-        12 => register_evm_version!(registry, cid, fil_actor_evm_state::v12),
-        13 => register_evm_version!(registry, cid, fil_actor_evm_state::v13),
-        14 => register_evm_version!(registry, cid, fil_actor_evm_state::v14),
-        15 => register_evm_version!(registry, cid, fil_actor_evm_state::v15),
-        16 => register_evm_version!(registry, cid, fil_actor_evm_state::v16),
-        _ => {}
+        ActorVersion::V8 | ActorVersion::V9 => {}
+        ActorVersion::V10 => register_evm_version!(registry, cid, fil_actor_evm_state::v10),
+        ActorVersion::V11 => register_evm_version!(registry, cid, fil_actor_evm_state::v11),
+        ActorVersion::V12 => register_evm_version!(registry, cid, fil_actor_evm_state::v12),
+        ActorVersion::V13 => register_evm_version!(registry, cid, fil_actor_evm_state::v13),
+        ActorVersion::V14 => register_evm_version!(registry, cid, fil_actor_evm_state::v14),
+        ActorVersion::V15 => register_evm_version!(registry, cid, fil_actor_evm_state::v15),
+        ActorVersion::V16 => register_evm_version!(registry, cid, fil_actor_evm_state::v16),
+        ActorVersion::V17 => register_evm_version!(registry, cid, fil_actor_evm_state::v17),
     }
 }